+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix using Sortable in scrolling container 
 - Fix lazy loading images if parent element is fully clipped
 - Fix clicking Slideshow with `draggable: false` no longer pauses autoplay
+- Fix toggling modal while transition is in progress
 
 ## 3.1.6 (June 19, 2019)